Resolved a very big issue with JDK.
- Not sure which was the actual fix, but it only worked until I added in the last part in build.gradle.

Here is a step-by-step of my troubleshooting, I used chatGPT for help:

Step 01: I went to 'File > Project Structure...'. Which is where I changed the SDK from 'Java 15' to 'temurin-1.8 Eclipse Temurin Version 1.8.0_392'.
     - This is the 1.8 JDK for Adoptium, which was formerly known as 'AdoptOpenJDK' and is now known as 'Temurin JDK'.
     - I also made sure to set the 'Language level' to '8 - Lambdas, type annotations etc.'.
     - All of this still failed. However, the 'Language level' kept getting reset to 'SDK Default' after a while as well. I left it at that, and it's still working now after the final fix.

Step 02: Another thing I tried was to go to 'File > Invalidate Caches', which gave me a popup with a bunch of optional checkboxes.
     - I did not select any of the options.
     - I clicked the button 'Invalidate and Restart'.
     - After the IDE restarted, I tried running the gradle, and it still didn't work

Step 03: In 'build.gradle' I added subprojects and put the repositories I had created before in it, along with a new snippet of code, which is the tasks.withType(JavaCompile) method.
     - Do note that this method was different before, it was 'Code1'. Which failed.
     - Then it was changed to 'Code2'. Which still failed.
     - Finally it was changed to 'Code5'. Which worked.

Step 04: In gradle.properties I added a new line of code: 'org.gradle.java.home=C:/Program Files/Eclipse Adoptium/jdk-8.0.392.8-hotspot'. Which still failed.

Step 05: Then in 'build.gradle' I added the 'task wrapper' as 'Code3', which gave me an error saying that it already existed, even though I couldn't find it when I searched for it using 'Crtl + R'.
     - I ran the command using './gradlew wrapper --gradle-version 7.3.3'.
     - Then I fixed the code to 'Code4'. Which still failed, but didn't give me the error that 'Code3' gave.

Step 06: Finally, in 'build.gradle', I changed 'Code2' to 'Code5'. This was what caused it to work when I tried running the gradle through both the terminal with './gradlew runClient' and once again through the 'Gradle > runClient' window on the top right.

Since I did so many things, I'm not sure which of the things really helped along with Step06 to make it work, but Step06 was the big one that "connected the dots" per se.

----------------------------------------------------
These are references and structured like methods to make it easier to read when explaining above.
----------------------------------------------------
Code1 {
     tasks.withType(JavaCompile) {
             options.fork = true
             options.forkOptions.executable = 'javac'
         }
}
----------------------------------------------------
Code2 {
    tasks.withType(JavaCompile) {
        options.fork = true
        options.forkOptions.executable = System.getenv('JAVA_HOME') + '/bin/javac'
    }
}
----------------------------------------------------
Code3 {
    task wrapper(type: Wrapper) {
        gradleVersion = '7.3.3' // Use the Gradle version you know is compatible
    }
}
----------------------------------------------------
Code4 {
    // Check if the 'wrapper' task is not already defined
    if (!project.tasks.findByName('wrapper')) {
        task wrapper(type: Wrapper) {
            gradleVersion = '7.3.3' // Use the Gradle version you know is compatible
        }
    }
}
----------------------------------------------------
Code5 {
    tasks.withType(JavaCompile) {
            options.fork = true
            options.forkOptions.executable = 'java'
            options.forkOptions.with {
                environment['JAVA_HOME'] = 'C:/Program Files/Eclipse Adoptium/jdk-8.0.392.8-hotspot'
            }
}
----------------------------------------------------